treat admin routes as child routes of main component
