+++ src/app/admin/admin-routing.module.ts
@@ -5,13 +5,20 @@ import { EditArticleComponent } from './edit-article/edit-article.component';
 import { EditFileComponent } from './edit-file/edit-file.component';
 import { ListFilesComponent } from './list-files/list-files.component';
 import { ListArticlesComponent } from './list-articles/list-articles.component';
+import { MainComponent } from './main/main.component';
 
 const routes: Routes = [
-  { path: '', component: HomeComponent },
-  { path: 'articles/:id', component: EditArticleComponent },
-  { path: 'files/:id', component: EditFileComponent },
-  { path: 'files', component: ListFilesComponent },
-  { path: 'articles', component: ListArticlesComponent },
+  {
+    path: '',
+    component: MainComponent,
+    children: [
+      { path: '', component: HomeComponent },
+      { path: 'articles/:id', component: EditArticleComponent },
+      { path: 'files/:id', component: EditFileComponent },
+      { path: 'files', component: ListFilesComponent },
+      { path: 'articles', component: ListArticlesComponent },
+    ]
+  },
 ];
 
 @NgModule({
