+++ src/app/admin/main/main.component.css
@@ -1,0 +1,12 @@
+.fill-space {
+    flex: 1 1 auto;
+}
+
+.home-link {
+    text-decoration: none;
+    color: white;
+}
+
+#userid {
+    font-size: small;
+}
