+++ src/app/admin/main/main.component.html
@@ -1,0 +1,28 @@
+<mat-toolbar color="primary">
+  <a class="home-link" routerLink="/">Fitchpork</a>
+  <span class="fill-space"></span>
+  <ng-container *ngIf="loggedIn | async">
+    <span id="userid">User: {{ (user | async)?.email }}</span>
+    <button mat-raised-button color="accent" (click)="logout()">Logout</button>
+  </ng-container>
+</mat-toolbar>
+<mat-sidenav-container>
+  <mat-sidenav-content>
+    <ng-container *ngIf="!(loggedIn | async)">
+      <mat-form-field>
+        <input type="email" matInput placeholder="Email" [(ngModel)]="userToLogIn.email">
+      </mat-form-field>
+      <mat-form-field>
+        <input type="password" matInput placeholder="Password" [(ngModel)]="userToLogIn.password">
+      </mat-form-field>
+      <div>
+        <button mat-raised-button (click)="login()">Login</button>
+        <div *ngIf="errorMessage">{{ errorMessage }}</div>
+      </div>
+    </ng-container>
+    <ng-container *ngIf="(loggedIn | async)">
+      <router-outlet></router-outlet>
+    </ng-container>
+    <br>
+  </mat-sidenav-content>
+</mat-sidenav-container>
