+++ src/app/app.component.spec.ts
@@ -13,19 +13,6 @@ import { FilesStorageService } from './files-storage.service';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 
 describe('AppComponent', () => {
-  const authenticationServiceStub = {
-    getAuthState: jasmine.createSpy('getAuthState'),
-    isLoggedIn: jasmine.createSpy('isLoggedIn').and.returnValue(Observable.of(true)),
-    login() { },
-    logout() { }
-  };
-
-  const articlesServiceStub = {
-    getArticles: jasmine.createSpy('getArticles').and.returnValue(Observable.of([])),
-  };
-
-  const imagesStorageServiceStub = {
-  };
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -37,11 +24,6 @@ describe('AppComponent', () => {
         AppComponent,
       ],
       schemas: [ CUSTOM_ELEMENTS_SCHEMA ],
-      providers: [
-        { provide: AuthenticationService, useValue: authenticationServiceStub },
-        { provide: ArticlesService, useValue: articlesServiceStub },
-        { provide: FilesStorageService, useValue: imagesStorageServiceStub },
-      ]
     }).compileComponents();
   }));
 
@@ -49,16 +31,5 @@ describe('AppComponent', () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
     expect(app).toBeTruthy();
-  }));
-
-  it('should render the user id', fakeAsync(() => {
-    authenticationServiceStub.getAuthState.and.returnValue(Observable.of({ email: 'test@test.de' }));
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    tick();
-    fixture.detectChanges();
-    const compiled = fixture.debugElement.nativeElement;
-    expect(authenticationServiceStub.getAuthState).toHaveBeenCalled();
-    expect(compiled.querySelector('#userid').textContent).toContain('User: test@test.de');
   }));
 });
